Extract InnerGameBoard CenterBar subcomponents
